+++ apps/fluent-tester/src/FluentTester/TestComponents/TextExperimental/consts.ts
@@ -1,0 +1,2 @@
+export const HOMEPAGE_EXPERIMENTAL_TEXT_BUTTON = 'Homepage_Text_Experimental_Button';
+export const TEXT_TESTPAGE = 'Text_Experimental_TestPage';